[io][svg] Add variant support for SVG writer.
Rename the file write_svg.hpp to write.hpp for consistency with other IOs.
Move dispatches for multi geometries to write.hpp.
Leave the old files for backward compatibility.
